requirements.txt
pyyaml
requests
elasticsearch
autogen-agentchat
autogen-ext[openai]
